fix: relocate interfaces file; lint stuff
