+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.ts",
   "scripts": {
     "start": "node ./dist/server/index.js",
-    "start-server": "tsc --project tsconfig.server.json && node ./dist/server/index.js | pino-pretty -c -t",
+    "start-server": "tsc --project tsconfig.server.json && node ./dist/index.js | pino-pretty -c -t",
     "start-react": "react-scripts start",
     "build-react": "react-scripts --openssl-legacy-provider build",
     "test": "yarn build-react && jest --coverage --watchAll=false",
