+++ server/interfaces/server.ts
