+++ tsconfig.json
@@ -25,7 +25,6 @@
     "noFallthroughCasesInSwitch": true
   },
   "include": [
-    "src",
-    "Interfaces"
+    "src"
   ]
 }
